move token functions to utils file
