+++ backend/package.json
@@ -14,6 +14,7 @@
 		"test": "jest"
 	},
 	"dependencies": {
+		"@dqbd/tiktoken": "^1.0.7",
 		"cors": "^2.8.5",
 		"d3-dsv": "^2.0.0",
 		"dotenv": "^16.3.1",
@@ -24,7 +25,8 @@
 		"openai": "^4.19.0",
 		"openai-chat-tokens": "^0.2.8",
 		"pdf-parse": "^1.1.1",
-		"react": "^18.2.0"
+		"react": "^18.2.0",
+		"tiktoken": "^1.0.11"
 	},
 	"devDependencies": {
 		"@types/cors": "^2.8.16",
